Timer, CSV export, XML feldolgozás készen! Belefektetett eddig: 11 óra.

+++ IRF_Project_GQOTXA/IRF_Project_GQOTXA/Nyitooldal.cs
@@ -20,7 +20,7 @@ namespace IRF_Project_GQOTXA
             InitializeComponent();
             labelEmail.Text = email;
             Random rnd = new Random();
-            int osszeg = (rnd.Next(1000, 50000)*100);
+            decimal osszeg = (rnd.Next(1000, 50000)*100);
             InvisibleLabel.Text = osszeg.ToString();
             InvisibleLabel.BackColor = Nyitooldal.DefaultBackColor;
             InvisibleLabel.ForeColor = Nyitooldal.DefaultBackColor;
